+++ pysparkling/sql/expressions/dates.py
@@ -3,8 +3,9 @@ import datetime
 from dateutil.relativedelta import relativedelta
 import pytz
 
+from .._date_time import get_unix_timestamp_parser
 from ...utils import parse_tz
-from ..casts import get_time_formatter, get_unix_timestamp_parser
+from ..casts import get_time_formatter
 from ..types import DateType, FloatType, TimestampType
 from .expressions import Expression, UnaryExpression
