+++ pysparkling/utils.py
@@ -1,24 +1,14 @@
-import collections
 import datetime
 import itertools
-import json
 import math
-from operator import itemgetter
 import random
 import re
 import sys
+from operator import itemgetter
 from typing import List, Optional, Union
 
 import pytz
 from pytz import UnknownTimeZoneError
-
-from .sql.casts import get_time_formatter
-from .sql.internal_utils.joins import (
-    CROSS_JOIN, FULL_JOIN, INNER_JOIN, LEFT_ANTI_JOIN, LEFT_JOIN, LEFT_SEMI_JOIN, RIGHT_JOIN
-)
-from .sql.schema_utils import get_on_fields
-from .sql.types import create_row, Row, row_from_keyed_values
-from .sql.utils import IllegalArgumentException
 
 
 class Tokenizer:
@@ -175,68 +165,6 @@ def get_keyfunc(cols, schema, nulls_are_smaller=False):
 
     return key
 
-
-FULL_WIDTH_REGEX = re.compile(
-    "["
-    + r"\u1100-\u115F"
-    + r"\u2E80-\uA4CF"
-    + r"\uAC00-\uD7A3"
-    + r"\uF900-\uFAFF"
-    + r"\uFE10-\uFE19"
-    + r"\uFE30-\uFE6F"
-    + r"\uFF00-\uFF60"
-    + r"\uFFE0-\uFFE6"
-    + "]"
-)
-
-
-def str_half_width(value):
-    """
-    Compute string length with full width characters counting for 2 normal ones
-    """
-    string = format_cell(value)
-    if string is None:
-        return 0
-    if not isinstance(string, str):
-        string = str(string)
-    return len(string) + len(FULL_WIDTH_REGEX.findall(string))
-
-
-def pad_cell(cell, truncate, col_width):
-    """
-    Compute how to pad the value "cell" truncated to truncate so that it fits col width
-    :param cell: Any
-    :param truncate: int
-    :param col_width: int
-    :return:
-    """
-    cell = format_cell(cell)
-    cell_width = col_width - str_half_width(cell) + len(cell)
-    if truncate > 0:
-        return cell.rjust(cell_width)
-    return cell.ljust(cell_width)
-
-
-def format_cell(value):
-    """
-    Convert a cell value to a string using the logic needed in DataFrame.show()
-    """
-    if value is None:
-        return "null"
-    if isinstance(value, bool):
-        return str(value).lower()
-    if isinstance(value, Row):
-        return f"[{', '.join(format_cell(sub_value) for sub_value in value)}]"
-    if isinstance(value, dict):
-        return "[{0}]".format(
-            ", ".join(
-                f"{format_cell(key)} -> {format_cell(sub_value)}"
-                for key, sub_value in value.items()
-            )
-        )
-    return str(value)
-
-
 class MonotonicallyIncreasingIDGenerator:
     def __init__(self, partition_index):
         self.value = partition_index * 8589934592 - 1
@@ -360,49 +288,6 @@ class MurmurHash3:
     @staticmethod
     def rotl(i, distance):
         return i << distance
-
-
-def merge_rows(left, right):
-    return create_row(
-        itertools.chain(left.__fields__, right.__fields__),
-        left + right
-    )
-
-
-def merge_rows_joined_on_values(left, right, left_schema, right_schema, how, on):
-    left_names = left_schema.names
-    right_names = right_schema.names
-
-    left_on_fields, right_on_fields = get_on_fields(left_schema, right_schema, on)
-
-    on_parts = [
-        (on_field, left[on_field] if left is not None else right[on_field])
-        for on_field in on
-    ]
-
-    if left is None and how in (FULL_JOIN, RIGHT_JOIN):
-        left = create_row(left_names, [None for _ in left_names])
-    if right is None and how in (LEFT_JOIN, FULL_JOIN):
-        right = create_row(right_names, [None for _ in right_names])
-
-    left_parts = (
-        (field.name, value)
-        for field, value in zip(left_schema.fields, left)
-        if field not in left_on_fields
-    )
-
-    if how in (INNER_JOIN, CROSS_JOIN, LEFT_JOIN, FULL_JOIN, RIGHT_JOIN):
-        right_parts = (
-            (field.name, value)
-            for field, value in zip(right_schema.fields, right)
-            if field not in right_on_fields
-        )
-    elif how in (LEFT_SEMI_JOIN, LEFT_ANTI_JOIN):
-        right_parts = ()
-    else:
-        raise IllegalArgumentException(f"Argument 'how' cannot be '{how}'")
-
-    return row_from_keyed_values(itertools.chain(on_parts, left_parts, right_parts))
 
 
 def strhash(string):
@@ -564,49 +449,3 @@ def levenshtein_distance(str1, str2):
         levenshtein_distance(str1[1:], str2) + 1,
         levenshtein_distance(str1, str2[1:]) + 1
     )
-
-
-def get_json_encoder(options):
-    """
-    Returns a JsonEncoder which convert Rows to json with the same behavior
-    and conversion logic as PySpark.
-
-    :param date_formatter: a function that convert a date into a string
-    :param timestamp_formatter: a function that convert a timestamp into a string
-    :return: type
-    """
-    date_format = options.get("dateformat", "yyyy-MM-dd")
-    timestamp_format = options.get("timestampformat", "yyyy-MM-dd'T'HH:mm:ss.SSSXXX")
-    date_formatter = get_time_formatter(date_format)
-    timestamp_formatter = get_time_formatter(timestamp_format)
-
-    class CustomJSONEncoder(json.JSONEncoder):
-        def encode(self, o):
-            def encode_rows(item):
-                if isinstance(item, Row):
-                    return collections.OrderedDict(
-                        (key, encode_rows(value))
-                        for key, value in zip(item.__fields__, item)
-                    )
-                if isinstance(item, (list, tuple)):
-                    return [encode_rows(e) for e in item]
-                if isinstance(item, dict):
-                    return collections.OrderedDict(
-                        (key, encode_rows(value))
-                        for key, value in item.items()
-                    )
-                return item
-
-            return super().encode(encode_rows(o))
-
-        # default can be overridden if passed a parameter during init
-        # pylint doesn't like the behavior but it is the expected one
-        # pylint: disable=E0202
-        def default(self, o):
-            if isinstance(o, datetime.datetime):
-                return timestamp_formatter(o)
-            if isinstance(o, datetime.date):
-                return date_formatter(o)
-            return super().default(o)
-
-    return CustomJSONEncoder
